Add files via upload
Data analysis project.  Data used is about Yelp reviews.  Project goal was to predict star ratings.  This was a group project done for a Linear Regression class.  Generally, our method was to search for the most commonly used words within the reviews and use those words as independent variables.  Additionally, we used interaction of the most commonly used words with words such as "not" or "but" because we believe that the presence of these words would change the general feeling of a review.